Implement beginnings of game state and actiond dispatching
Implemented PlayerDescriptor, RulesAuthority, DefaultRulesAuthority and some other
items that should prove decent.

The system is based around keeping an immutable state object inside the game. Every
time an action is taken by a player, the RulesAuthority creates a new state for the
game to use, which should be wholly representative of the game.